Renamed Curl to Connection and moved to dedicated namespace
